+++ packages/backend/src/app.ts
@@ -1,5 +1,5 @@
 // For more information about this file see https://dove.feathersjs.com/guides/cli/application.html
-import { FeathersService, ServiceInterface, ServiceOptions, feathers } from '@feathersjs/feathers';
+import { feathers } from '@feathersjs/feathers';
 import configuration from '@feathersjs/configuration';
 import {
   koa,
@@ -12,8 +12,6 @@ import {
 } from '@feathersjs/koa';
 import socketio from '@feathersjs/socketio';
 import { feathersCasl } from 'feathers-casl';
-
-import type { Application, ServiceTypes } from './declarations';
 import { mongodb } from './mongodb';
 import { authentication } from './authentication';
 import { services } from './services/index';
@@ -21,51 +19,10 @@ import { channels } from './channels';
 import { getRegisteredServices } from './utils/registered-services';
 import appHooks from './app.hooks';
 
-function createApp() {
-  const a = koa(feathers()) as any;
+const app = koa(feathers()) as any;
 
-  // Load our app configuration (see config/ folder)
-  a.configure(configuration());
-
-  let apiPrefix = ((a.get('apiPrefix') as string) || '')
-    .split('/')
-    .filter((x) => x !== '')
-    .join('/');
-  if (apiPrefix.length > 0) {
-    apiPrefix += '/';
-  }
-  a.set('apiPrefix', apiPrefix);
-
-  if (a.get('authentication').enabled) {
-    a.get('authentication').service = apiPrefix + a.get('authentication').service;
-  }
-
-  a.getService = function <L extends keyof ServiceTypes>(
-    location: L,
-  ): FeathersService<Application, ServiceTypes[L]> {
-    return a.service(apiPrefix + location);
-  };
-
-  a.getServicePath = function <L extends keyof ServiceTypes>(location: L): string {
-    return apiPrefix + location;
-  };
-
-  a.declareService = function <L extends keyof ServiceTypes>(
-    location: L,
-    service: keyof any extends keyof ServiceTypes
-      ? ServiceInterface | Application
-      : ServiceTypes[L],
-    options?: ServiceOptions<keyof any extends keyof ServiceTypes ? string : keyof ServiceTypes[L]>,
-  ) {
-    a.use(apiPrefix + location, service, options);
-
-    return (a as Application).getService(location);
-  };
-
-  return a as Application;
-}
-
-const app = createApp();
+// Load our app configuration (see config/ folder)
+app.configure(configuration());
 
 // Set up Koa middleware
 app.use(cors());