+++ packages/backend/src/services/info/info.ts
@@ -11,14 +11,14 @@ export * from './info.class';
 // A configure function that registers the service and its hooks via `app.configure`
 export const info = (app: Application) => {
   // Register our service on the Feathers application
-  app.declareService(infoPath, new InfoService(getOptions(app)), {
+  app.use(infoPath, new InfoService(getOptions(app)), {
     // A list of all methods this service exposes externally
     methods: infoMethods,
     // You can add additional custom events to be sent to clients here
     events: [],
   });
   // Initialize hooks
-  app.getService(infoPath).hooks({
+  app.service(infoPath).hooks({
     around: {
       all: [],
     },