+++ packages/backend/test/services/info/info.test.ts
@@ -4,7 +4,7 @@ import { app } from '../../../src/app';
 
 describe('info service', () => {
   it('registered the service', () => {
-    const service = app.getService('info');
+    const service = app.service('info');
 
     assert.ok(service, 'Registered the service');
   });