+++ packages/backend/test/services/users/users.test.ts
@@ -4,7 +4,7 @@ import { app } from '../../../src/app';
 
 describe('users service', () => {
   it('registered the service', () => {
-    const service = app.getService('users');
+    const service = app.service('users');
 
     assert.ok(service, 'Registered the service');
   });