+++ packages/core/src/components/media-recorder/index.ts
@@ -1,0 +1,7 @@
+import { MediaRecorder } from './media-recorder.component';
+
+export function mediaRecorder(...args: ConstructorParameters<typeof MediaRecorder>): MediaRecorder {
+  return new MediaRecorder(...args);
+}
+
+export type { MediaRecorder };