+++ apps/demos/CHANGELOG.md
@@ -1,5 +1,27 @@
 # demos
 
+## 0.6.5
+
+### Patch Changes
+
+- @marcellejs/core@0.6.5
+
+## 0.6.4
+
+### Patch Changes
+
+- Updated dependencies [1dd0d82]
+  - @marcellejs/core@0.6.4
+
+## 0.6.3
+
+### Patch Changes
+
+- Updated dependencies [6073dab]
+- Updated dependencies [c9a770e]
+- Updated dependencies [c9a770e]
+  - @marcellejs/core@0.6.3
+
 ## 0.6.2
 
 ### Patch Changes